pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interview-phrase-tagger"
version = "0.0.1"
description = "A tool to tag phrases in interview transcripts."
authors = [{ name = "Juan Rojas", email = "jdrojasga@unal.edu.co" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.12.0",
    "typer>=0.20.0",
    "transformers>=4.36.0,<5.0.0",
    "torch>=2.1.0,<2.3.0",
    "numpy<2.0",
    "pyyaml>=6.0",
]

# Expose the CLI command `iptag`
[project.scripts]
iptag = "iptag.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

# -----------------------------
# Ruff configuration
# -----------------------------
[tool.ruff]
# Allow lines to be as long as 88.
line-length = 88

[tool.ruff.lint]
select = [
    # pydocstyle
    "D",
    # pycodestyle
    "E",
    # Pyflakes 
    "F",
    # isort
    "I",
    # pandas-vet
    "PD",
]

pydocstyle.convention = "google"

# -----------------------------
# Dev dependencies
# -----------------------------
[dependency-groups]
dev = ["pytest>=9.0.1", "pytest-cov>=6.0.0", "ruff>=0.11.8"]

# -----------------------------
# Pytest configuration
# -----------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--strict-markers", "--strict-config", "-ra"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
